requirements.txt
argparse==1.4.0
python-logstash==0.4.6
python-logstash-async